feat(frontend): extract FoodOffersScroll list
