+++ apps/frontend/app/components/FoodOffersScrollList/styles.ts
@@ -1,0 +1,27 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+  dayContainer: {
+    padding: 10,
+  },
+  dateHeader: {
+    fontSize: 18,
+    marginBottom: 8,
+  },
+  loader: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  foodContainer: {
+    width: '100%',
+    flexDirection: 'row',
+    alignItems: 'stretch',
+    flexWrap: 'wrap',
+    marginTop: 20,
+  },
+  feebackContainer: {
+    width: '100%',
+    marginTop: 20,
+  },
+});
